requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http>=5.5.0
apache-airflow-providers-postgres>=5.5.0
requests
